move hd void* casts into a utility function to clean up renderData.cpp
